# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Architectural drawing management API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in to the drawing portal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed-in portal user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortalUser"
        "401":
          description: Invalid credentials
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List employee attendance for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Employee attendance register
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEntry"
  /attendance/me:
    get:
      operationId: getMyAttendance
      tags: [attendance]
      summary: Get the signed-in employee's attendance for a date
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Signed-in employee attendance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEntry"
  /attendance/me/month:
    get:
      operationId: getMyAttendanceMonth
      tags: [attendance]
      summary: Get the signed-in employee's attendance for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
            pattern: "^[0-9]{4}-(0[1-9]|1[0-2])$"
      responses:
        "200":
          description: Signed-in employee monthly attendance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Invalid or future month
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /attendance/self-checkin:
    post:
      operationId: selfCheckinAttendance
      tags: [attendance]
      summary: Check in for the signed-in employee with location evidence
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceSelfCheckinInput"
      responses:
        "201":
          description: Self-check-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Invalid location evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Attendance is already recorded for the day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /attendance/{employeeId}:
    put:
      operationId: recordAttendance
      tags: [attendance]
      summary: Correct an existing location-backed attendance check-in
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: number
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "200":
          description: Recorded attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEntry"
        "400":
          description: Invalid attendance record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Employee not found
...
  /categories:
    get:
      operationId: listCategories
      tags: [drawings]
      summary: List drawing categories
      responses:
        "200":
          description: Drawing categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [drawings]
      summary: Add a drawing category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [drawings]
      summary: Rename a drawing category
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [drawings]
      summary: Delete a drawing category
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: number
        minimum: 1
...
  schemas:
...
    AttendanceStatus:
      type: string
      enum: [present, late, absent, leave, remote]
    AttendanceEntry:
      type: object
      properties:
        employeeId:
          type: number
        employeeName:
          type: string
        username:
          type: ["string", "null"]
        attendanceDate:
          type: string
          format: date
        status:
          type: ["string", "null"]
          enum: [present, late, absent, leave, remote, null]
        reason:
          type: ["string", "null"]
        recordedBy:
          type: ["number", "null"]
        recordedAt:
          type: ["string", "null"]
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        accuracyMeters:
          type: ["number", "null"]
        selfCheckinAt:
          type: ["string", "null"]
          format: date-time
      required: [employeeId, employeeName, username, attendanceDate, status, reason, recordedBy, recordedAt, updatedAt, latitude, longitude, accuracyMeters, selfCheckinAt]
    AttendanceInput:
      type: object
      properties:
        attendanceDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/AttendanceStatus"
        reason:
          type: string
      required: [attendanceDate, status]
    AttendanceSelfCheckinInput:
      type: object
      properties:
        attendanceDate:
          type: string
          format: date
        workLocation:
          type: string
          enum: [office, remote]
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        accuracyMeters:
          type: number
          minimum: 0
          maximum: 100000
      required: [attendanceDate, workLocation, latitude, longitude, accuracyMeters]
    PortalUser:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        username:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, username, role, active, createdAt]
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
...
    Category:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
      required: [name]
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]